GCW-3601 Refactor donors test
