--- src/test-utils/matchers/index.ts
@@ -1,13 +1,0 @@
-const expectToBeOnPage = (
-  container: HTMLElement,
-  myPath: string,
-  expectedPage: string
-) => {
-  const expectedPath = `/${expectedPage}`;
-  expect(myPath).toEqual(expectedPath);
-  expect(container.querySelector("ion-title")).toHaveTextContent(
-    new RegExp(expectedPage, "i")
-  );
-};
-
-export { expectToBeOnPage };
